fix(clippy): clear strict lints blocking CI on the reflection slice
- reflection.rs: use size_of_val for the material slice
- reflection.rs: replace !(exit > 0.) with exit <= 0.; the preceding
  is_finite guard short-circuits NaN, so the rewrite is exact
- reflection_validation.rs: mark the unused harness parameter
- reflection_check.rs: drop a redundant f64 -> f64 cast

These were masked in CI because the earlier rustfmt failure aborted the
step before clippy ran.

+++ apps/explorer/src/reflection_check.rs
@@ -483,7 +483,7 @@ impl ApplicationHandler for ReflectionCheck {
                 m.intervals.push(interval);
                 m.calls.push(call_ms);
                 if let Some(timing) = self.renderer.as_ref().unwrap().gpu_timings() {
-                    m.gpu = Some(timing.render_ms as f64);
+                    m.gpu = Some(timing.render_ms);
                 }
             }
             if m.frames >= phase.measured && phase.measured > 0 {
